fix: Adjusts in remove method in users repository

+++ server/apps/user/src/application/use-cases/remove-user/remove-user.ts
@@ -1,11 +1,20 @@
+import { UserProps } from "../../entities/user";
 import { UserRepository } from "../../repositories/users-repository";
+
+type User = UserProps
 
 export class RemoveUser {
   constructor(
     private usersRepository: UserRepository
   ){}
 
-  async execute(id: string): Promise<void> {
+  async execute(id: string): Promise<User[]> {
+    const user = await this.usersRepository.findUnique(id);
+
+    if (!user) {
+      throw new Error('User does not exist')
+    }
+
     return await this.usersRepository.remove(id)
   }
 }
